Bugfix: Configuration of sensormode will not longer be reseted. Update on German Language strings.xml
