+++ impl/androidApp/api/src/main/java/org/mindroid/impl/configuration/RobotPortConfig.java
@@ -1,4 +1,4 @@
-package org.mindroid.android.app.SchuelerProjekt;
+package org.mindroid.impl.configuration;
 
 import org.mindroid.api.robot.IRobotPortConfig;
 import org.mindroid.common.messages.Motors;
