+++ impl/androidApp/app/src/main/res/values-b+de+DE/strings.xml
@@ -2,29 +2,49 @@
     <string name="app_name">Robodancer</string>
     <string name="navigation_drawer_open">Open navigation drawer</string>
     <string name="navigation_drawer_close">Close navigation drawer</string>
-    <string name="action_settings">Settings</string>
+    <string name="action_settings">Einstellungen</string>
 
     <!-- Activity titles -->
     <string name="title_activity_main">Mindroid</string>
 
-    <string name="dialog_msg_connect_to_robot">Mit dem Roboter verbinden</string>
-    <string name="btn_text_start_robot">Start</string>
+    <!-- Home Fragment -->
+    <string name="btn_text_activate_tethering">Aktiviere Tethering</string>
     <string name="btn_text_init_config">Initialisiere Konfiguration</string>
-    <string name="titleStatus">Status:</string>
-    <string name="not_running">Nicht aktiv</string>
-    <string name="connected">Verbindung:</string>
-    <string name="robot">Robot:</string>
-    <string name="btn_text_stop_robot">Stop</string>
-    <string name="not_connected">nicht verbunden</string>
-    <string name="title_activity_settings">SettingsActivity</string>
+    <string name="btn_text_start_robot">Start</string>
+    <string name="btn_text_stop_robot">stop</string>
+    <string name="btn_text_disconnect">Verbindung trennen</string>
+    <string name="btn_text_connect">Verbindung zum Roboter herstellen</string>
+    <string name="dialog_msg_connect_to_robot">Verbindung wird aufgebaut..</string>
+    <string name="dialog_msg_init_config">Der Roboter wird konfiguriert..</string>
+    <string name="dialog_msg_disconnect_from_robot">Trennen..</string>
+    <string name="dialog_msg_start_robot">Roboter wird gestartet..</string>
+    <string name="dialog_msg_stop_robot">Roboter wird gestoppt..</string>
+
+    <!-- Settings Fragment -->
+    <string name="btn_text_save_settings">Einstellungen speichern</string>
+
+
+    <!-- Hardware Selection Fragment -->
+    <string name="view_sensortype">Sensor</string>
+    <string name="view_sensormode">Modus</string>
+    <string name="view_motortype">Motor</string>
+
+    <!-- RobotSetupInfo Fragment-->
+    <string name="tab_connection_tag">Info</string>
+
+    <!-- Configuration Fragment -->
+    <string name="tab_sensors_tag">Sensoren</string>
+    <string name="tab_motors_tag">Motoren</string>
+
+    <!-- Settings Fragment -->
+    <string name="msg_toast_settings_invalid_input">Ungueltige Eingabe!</string>
+    <string name="msg_taost_settings_saved">Einstellungen gespeichert!</string>
 
     <!-- Shared Preferences -->
     <string name="shared_pref_connection_Data">ConnectionData.xml</string>
+    <string name="shared_pref_portConfiguration">PortConfiguration.xml</string>
 
-    <!-- Settings Fragment -->
-    <string name="msg_taost_settings_saved">gespeichert</string>
-
-    <!-- DEFAULT VALUES -->
+    <!-- Shared Preferences: DEFAULT VALUES -->
     <string name="DEFAULT_EV3_BRICK_IP">192.168.0.240</string>
     <string name="DEFAULT_EV3_BRICK_PORT">33690</string>
     <string name="DEFAULT_MSG_SERVER_IP">192.168.188.38</string>
@@ -34,7 +54,7 @@
     <string name="DEFAULT_GROUP_ID">No groupID set</string>
 
 
-    <!-- KEYS for Shared Preferences - ConnectionData.xml -->
+    <!-- Shared Preferences: KEYS for Shared Preferences - ConnectionData.xml -->
     <string name="KEY_ROBOT_ID">ROBOT_ID</string>
     <string name="KEY_GROUP_ID">GROUP_ID</string>
     <string name="KEY_EV3_IP">KEY_EV3_IP</string>
@@ -43,15 +63,32 @@
     <string name="KEY_SERVER_TCP_PORT">KEY_SERVER_TCP_PORT</string>
     <string name="KEY_ROBOT_SERVER_TCP_PORT">KEY_ROBOT_SERVER_TCP_PORT</string>
 
+    <string name="KEY_SENSOR_S1">sensorS1</string>
+    <string name="KEY_SENSOR_S2">sensorS2</string>
+    <string name="KEY_SENSOR_S3">sensorS3</string>
+    <string name="KEY_SENSOR_S4">sensorS4</string>
+    <string name="KEY_SENSORMODE_S1">sensormodeS1</string>
+    <string name="KEY_SENSORMODE_S2">sensormodeS2</string>
+    <string name="KEY_SENSORMODE_S3">sensormodeS3</string>
+    <string name="KEY_SENSORMODE_S4">sensormodeS4</string>
+    <string name="KEY_MOTOR_A">motorA</string>
+    <string name="KEY_MOTOR_B">motorB</string>
+    <string name="KEY_MOTOR_C">motorC</string>
+    <string name="KEY_MOTOR_D">motorD</string>
+
+
+    <!-- ErrorDialog -->
+    <string name="text_ok_button">OK</string>
+
+    <!-- Navigation Drawer item titles -->
     <string name="title_section1">Home</string>
-    <string name="title_section2">Mein Robotor</string>
+    <string name="title_section2">Mein Roboter</string>
     <string name="title_section3">Einstellungen</string>
 
     <string name="action_example">Example action</string>
 
     <!-- TODO: Remove or change this placeholder text -->
     <string name="hello_blank_fragment">Hello blank fragment</string>
-    <string name="dialog_msg_init_config" />
-    <string name="dialog_msg_start_robot" />
-    <string name="dialog_msg_stop_robot" />
+
+
 </resources>
